+++ apps/backend/src/modules/projects/project.service.ts
@@ -1,8 +1,18 @@
-import { ExceptionMessage } from "~/libs/enums/enums.js";
+import { EMPTY_LENGTH } from "~/libs/constants/constants.js";
+import {
+	ExceptionMessage,
+	PermissionKey,
+	ProjectPermissionKey,
+} from "~/libs/enums/enums.js";
+import { getDifferenceInDays } from "~/libs/helpers/helpers.js";
 import { HTTPCode } from "~/libs/modules/http/http.js";
+import { type Logger } from "~/libs/modules/logger/logger.js";
 import { type Service } from "~/libs/types/types.js";
 import { type ProjectApiKeyService } from "~/modules/project-api-keys/project-api-key.service.js";
 
+import { type NotificationService } from "../notifications/notification.service.js";
+import { type UserService } from "../users/user.service.js";
+import { NOTIFICATION_DAY_THRESHOLD } from "./libs/constants/constants.js";
 import { ProjectError } from "./libs/exceptions/exceptions.js";
 import {
 	type ProjectCreateRequestDto,
@@ -16,15 +26,35 @@ import {
 import { ProjectEntity } from "./project.entity.js";
 import { type ProjectRepository } from "./project.repository.js";
 
+type Constructor = {
+	logger: Logger;
+	notificationService: NotificationService;
+	projectApiKeyService: ProjectApiKeyService;
+	projectRepository: ProjectRepository;
+	userService: UserService;
+};
+
 class ProjectService implements Service {
+	private logger: Logger;
+
+	private notificationService: NotificationService;
+
 	private projectApiKeyService: ProjectApiKeyService;
 
 	private projectRepository: ProjectRepository;
 
-	public constructor(
-		projectRepository: ProjectRepository,
-		projectApiKeyService: ProjectApiKeyService,
-	) {
+	private userService: UserService;
+
+	public constructor({
+		logger,
+		notificationService,
+		projectApiKeyService,
+		projectRepository,
+		userService,
+	}: Constructor) {
+		this.logger = logger;
+		this.notificationService = notificationService;
+		this.userService = userService;
 		this.projectRepository = projectRepository;
 		this.projectApiKeyService = projectApiKeyService;
 	}
@@ -105,6 +135,19 @@ class ProjectService implements Service {
 		};
 	}
 
+	public async findInactiveProjects(
+		thresholdInDays: number,
+	): Promise<ProjectGetAllItemResponseDto[]> {
+		const projects =
+			await this.projectRepository.findInactiveProjects(thresholdInDays);
+
+		return projects.map((project) => {
+			const { id, lastActivityDate, name } = project.toObject();
+
+			return { id, lastActivityDate, name };
+		});
+	}
+
 	public async patch(
 		id: number,
 		projectData: ProjectPatchRequestDto,
@@ -143,6 +186,58 @@ class ProjectService implements Service {
 			...project,
 			apiKey,
 		};
+	}
+
+	public async processInactiveProjects(): Promise<void> {
+		const [firstThreshold, secondThreshold] = NOTIFICATION_DAY_THRESHOLD;
+		const projects = await this.findInactiveProjects(firstThreshold);
+
+		for (const project of projects) {
+			const daysInactive = getDifferenceInDays(
+				new Date(),
+				new Date(project.lastActivityDate as string),
+			);
+
+			if (
+				daysInactive === firstThreshold ||
+				daysInactive % secondThreshold === EMPTY_LENGTH
+			) {
+				const projectUsers =
+					await this.userService.findAllWithProjectPermissions(
+						[
+							ProjectPermissionKey.EDIT_PROJECT,
+							ProjectPermissionKey.VIEW_PROJECT,
+							ProjectPermissionKey.MANAGE_PROJECT,
+						],
+						project.id,
+					);
+
+				const rootUsers = await this.userService.findAllWithRootPermissions([
+					PermissionKey.MANAGE_ALL_PROJECTS,
+					PermissionKey.VIEW_ALL_PROJECTS,
+				]);
+
+				const filteredRootUsers = rootUsers.filter(
+					(rootUser) =>
+						!projectUsers.some((projectUser) => projectUser.id === rootUser.id),
+				);
+
+				const users = [...projectUsers, ...filteredRootUsers];
+
+				if (users.length > EMPTY_LENGTH) {
+					const userIds = users.map((user) => user.id);
+
+					await this.notificationService.bulkCreate({
+						payload: `Statistics for ${project.name} is not uploaded for ${daysInactive.toString()} days.`,
+						receiverUserIds: userIds,
+					});
+
+					this.logger.info(
+						`Sent notifications to users for project ${project.name}.`,
+					);
+				}
+			}
+		}
 	}
 
 	public update(): ReturnType<Service["update"]> {