+++ apps/frontend/src/libs/enums/notification-message.enum.ts
@@ -1,4 +1,5 @@
 const NotificationMessage = {
+	CONTRIBUTOR_UPDATE_SUCCESS: "Contributor was successfully updated.",
 	GROUP_CREATE_SUCCESS: "Group was successfully created.",
 	GROUP_DELETE_SUCCESS: "Group was successfully deleted.",
 	GROUP_UPDATE_SUCCESS: "Group was successfully updated.",
@@ -11,6 +12,7 @@ const NotificationMessage = {
 	PROJECT_GROUP_DELETE_SUCCESS: "Project group was successfully deleted.",
 	PROJECT_GROUP_UPDATE_SUCCESS: "Project group was successfully updated.",
 	PROJECT_UPDATE_SUCCESS: "Project was successfully updated.",
+	SCRIPT_COPY_SUCCESS: "Script was successfully copied.",
 	USER_DELETE_SUCCESS: "User deleted successfully.",
 } as const;