+++ apps/frontend/src/pages/access-management/libs/helpers/get-user-columns/get-user-columns.helper.ts
@@ -1,0 +1,27 @@
+import { formatDate } from "~/libs/helpers/helpers.js";
+import { type TableColumn } from "~/libs/types/types.js";
+
+import { type UserRow } from "../../types/types.js";
+
+const getUserColumns = (): TableColumn<UserRow>[] => {
+	return [
+		{
+			accessorKey: "name",
+			header: "Name",
+			size: 300,
+		},
+		{
+			accessorFn: (user: UserRow): string => user.groups.join(", "),
+			header: "Groups",
+			size: 460,
+		},
+		{
+			accessorFn: (user: UserRow): string =>
+				formatDate(new Date(user.createdAt), "d MMM yyyy HH:mm"),
+			header: "Created At",
+			size: 200,
+		},
+	];
+};
+
+export { getUserColumns };