+++ packages/shared/src/libs/enums/enums.ts
@@ -3,5 +3,6 @@ export { AppEnvironment } from "./app-environment.enum.js";
 export { ContentType } from "./content-type.enum.js";
 export { ExceptionMessage } from "./exception-message.enum.js";
 export { PermissionKey } from "./permission-key.enum.js";
+export { ProjectPermissionKey } from "./project-permission-key.enum.js";
 export { ServerErrorType } from "./server-error-type.enum.js";
 export { SortType } from "./sort-type.enum.js";